feat: Add error handling to marketplace action modals
