+++ src/components/common/modals/TransferToken.tsx
@@ -24,10 +24,10 @@ import {
   UseDisclosureReturn
 } from '@chakra-ui/react';
 import { CheckCircle, AlertCircle, X, Plus } from 'react-feather';
-import { MinterButton } from '../common';
-import { useSelector, useDispatch } from '../../reducer';
-import { transferTokenAction } from '../../reducer/async/actions';
-import { clearError, setStatus, Status } from '../../reducer/slices/status';
+import { MinterButton } from '../../common/index';
+import { useSelector, useDispatch } from '../../../reducer';
+import { transferTokenAction } from '../../../reducer/async/actions';
+import { clearError, setStatus, Status } from '../../../reducer/slices/status';
 
 interface FormProps {
   initialRef: MutableRefObject<null>;
